requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==20.8b1
boxsdk==2.9.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.0
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.0
pycparser==2.20
PyJWT==1.7.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wrapt==1.12.1
